fix cvar check "warning"

+++ check-cvars.sh
@@ -6,16 +6,20 @@ check_files()
 {
 	countw=`awk ''"$3"' { print $2; }' "$1" | sort -u | tr -d '\r' | git hash-object --stdin | cut -c 1-32`
 	for b in $2; do
+		if [ "$b" = "balance-testing.cfg" ] || [ "$b" = "balance-testingxpm.cfg" ]; then
+			continue
+		fi
+
 		countb=`awk ''"$3"' { print $2; }' "$b" | sort -u | tr -d '\r' | git hash-object --stdin | cut -c 1-32`
 		if [ "$countw" != "$countb" ]; then
-			echo "Mismatch between "$1" and $b. Aborting."
+			echo "Mismatch between $1 and $b. Aborting."
 			echo "Differences are:"
 			echo "< missing in $b"
 			echo "> must get removed from $b"
 			A=`mktemp || echo a.tmp`
 			B=`mktemp || echo b.tmp`
-			awk ''"$3"' { print $2; }' "$1"      | sort -u | tr -d '\r' > "$A"
-			awk ''"$3"' { print $2; }' "$b"                     | sort -u | tr -d '\r' > "$B"
+			awk ''"$3"' { print $2; }' "$1" | sort -u | tr -d '\r' > "$A"
+			awk ''"$3"' { print $2; }' "$b" | sort -u | tr -d '\r' > "$B"
 			diff "$A" "$B" | grep '^[<>]' | sort
 			rm -f "$A" "$B"
 			errord=true
